requirements.txt
phidata
Groq
yfinance
duckduckgo-search